moved units module into nineml root package
