+++ nineml/document.py
@@ -186,9 +186,8 @@ class Document(dict, BaseNineMLObject):
             *[o.all_dimensions for o in self.itervalues()]))
         # Delete unused units from the document
         for k, o in self.items():
-            if ((isinstance(o, nineml.abstraction_layer.Unit) and
-                 o not in all_units) or
-                (isinstance(o, nineml.abstraction_layer.Dimension) and
+            if ((isinstance(o, nineml.Unit) and o not in all_units) or
+                (isinstance(o, nineml.Dimension) and
                  o not in all_dimensions)):
                 del self[k]
         # Add missing units and dimensions to the document
@@ -262,15 +261,11 @@ class Document(dict, BaseNineMLObject):
                     annotations = Annotations.from_xml(child)
                     continue
                 try:
-                    child_cls = getattr(nineml.user_layer, element_name)
+                    child_cls = getattr(nineml, element_name)
                 except AttributeError:
-                    try:
-                        child_cls = getattr(nineml.abstraction_layer,
-                                            element_name)
-                    except AttributeError:
-                        raise NineMLRuntimeError(
-                            "Did not find matching NineML class for '{}' "
-                            "element".format(element_name))
+                    raise NineMLRuntimeError(
+                        "Did not find matching NineML class for '{}' "
+                        "element".format(element_name))
                 if not issubclass(child_cls, DocumentLevelObject):
                     raise NineMLRuntimeError(
                         "'{}' is not a valid top-level NineML element"
