+++ test/unit/nineml_test/abstraction_layer/validators/dimensionality_test.py
@@ -2,7 +2,7 @@ import unittest
 from nineml.abstraction_layer import (
     Parameter, Constant, DynamicsClass, Regime, On, OutputEvent, StateVariable)
 from nineml.abstraction_layer.ports import AnalogSendPort, AnalogReceivePort
-from nineml.abstraction_layer import units as un
+from nineml import units as un
 
 
 class Dimensionality_test(unittest.TestCase):
@@ -12,16 +12,16 @@ class Dimensionality_test(unittest.TestCase):
         DynamicsClass(
             name='A',
             aliases=['A1:=P1 * SV2', 'A2 := ARP1 + SV2', 'A3 := SV1'],
+            state_variables=[
+                StateVariable('SV1', dimension=un.voltage),
+                StateVariable('SV2', dimension=un.current)],
             regimes=[
                 Regime(
                     'dSV1/dt = -SV1 / P1',
                     'dSV2/dt = A3 / ARP2 + SV2 / P2',
                     transitions=[On('SV1 > P3', do=[OutputEvent('emit')]),
                                  On('spikein', do=[OutputEvent('emit')])],
-                    name='R1',
-                    state_variables=[
-                        StateVariable('SV1', dimension=un.voltage),
-                        StateVariable('SV2', dimension=un.current)],
+                    name='R1'
                 ),
                 Regime(name='R2', transitions=On('SV1 > C1', to='R1'))
             ],
