+++ test/unit/test_userlayer.py
@@ -6,7 +6,7 @@ Tests for the user_layer module
 import unittest
 from lxml import etree
 from nineml.user_layer import Property
-from nineml.abstraction_layer import Unit, Dimension
+from nineml import Unit, Dimension
 from nineml.document import Document
 
 voltage = Dimension('voltage', m=1, l=2, t=-3, i=-1)
